// src/lib/enums.generated.ts
export const LoanType = {
  DSCR: "dscr",
  FIX_AND_FLIP: "fix_and_flip",
  GROUND_UP: "ground_up",
  BRIDGE: "bridge",
  PORTFOLIO: "portfolio",
  CASH_OUT_REFI: "cash_out_refi"
} as const;
export type LoanType = (typeof LoanType)[keyof typeof LoanType];